# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Energy and Water",
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/pregunta_frecuente": {
            "get": {
                "summary": "Obtener todas las preguntas frecuentes",
                "operationId": "getAllPreguntasFrecuentes",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "No se encontraron datos"
                    }
                }
            },
            "post": {
                "summary": "Agregar nueva pregunta frecuente",
                "operationId": "createPreguntaFrecuente",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pregunta",
                                    "respuesta",
                                    "activo"
                                ],
                                "properties": {
                                    "pregunta": {
                                        "type": "string"
                                    },
                                    "respuesta": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pregunta frecuente creada con éxito"
                    },
                    "400": {
                        "description": "Validación fallida"
                    }
                }
            }
        },
        "/v1/pregunta_frecuente/{id}": {
            "get": {
                "summary": "Obtener una pregunta frecuente por ID",
                "operationId": "getPreguntaFrecuenteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pregunta frecuente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Pregunta frecuente no existe"
                    }
                }
            },
            "put": {
                "summary": "Actualizar o crear una pregunta frecuente por ID",
                "operationId": "updatePreguntaFrecuente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pregunta frecuente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pregunta",
                                    "respuesta",
                                    "activo"
                                ],
                                "properties": {
                                    "pregunta": {
                                        "type": "string"
                                    },
                                    "respuesta": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pregunta frecuente actualizada con éxito"
                    },
                    "201": {
                        "description": "Pregunta frecuente creada con éxito"
                    },
                    "404": {
                        "description": "Pregunta frecuente no encontrada"
                    }
                }
            },
            "delete": {
                "summary": "Eliminar una pregunta frecuente por ID",
                "operationId": "deletePreguntaFrecuente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pregunta frecuente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pregunta frecuente eliminada con éxito"
                    },
                    "404": {
                        "description": "Pregunta frecuente no encontrada"
                    }
                }
            },
            "patch": {
                "summary": "Actualizar parcialmente una pregunta frecuente por ID",
                "operationId": "partialUpdatePreguntaFrecuente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pregunta frecuente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pregunta": {
                                        "type": "string"
                                    },
                                    "respuesta": {
                                        "type": "string"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pregunta frecuente actualizada con éxito"
                    },
                    "404": {
                        "description": "Pregunta frecuente no encontrada"
                    },
                    "400": {
                        "description": "No hay datos para actualizar"
                    }
                }
            }
        }
    }
}